Conflicts:
	src/com.bionic.book-store/web/js/admin_topic_list.js

+++ src/com.bionic.book-store/java/util/MultipartRequestMap.java
@@ -93,6 +93,7 @@ public class MultipartRequestMap extends HashMap<String, List<Object>> {
     }
 
     private void processFilePart(Part part, String fileName) throws IOException {
+        if (fileName.length() == 0) return;
         File tempFile = new File(STORAGE, fileName);
         tempFile.createNewFile();
         tempFile.deleteOnExit();
